fastAPI_InscribeME/requirements.txt
annotated-types==0.6.0
click==8.1.7
fastapi==0.110.0
gunicorn==21.2.0
mysql-connector-python==8.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-multipart==0.0.9
starlette==0.36.3
uvicorn==0.29.0
PyJWT==2.8.0